+++ spec/controllers/wallaby/application_controller_spec.rb
@@ -1,6 +1,16 @@
 require 'rails_helper'
 
 describe Wallaby::ApplicationController do
+  describe '#healthy' do
+    it 'returns healthy' do
+      get :healthy
+      expect(response.body).to eq 'healthy'
+
+      get :healthy, format: :json
+      expect(response.body).to eq 'healthy'
+    end
+  end
+
   describe 'error handling' do
     describe 'Wallaby::ResourceNotFound' do
       controller do