+++ spec/rails_helper.rb
@@ -1,11 +1,3 @@
-require 'simplecov'
-SimpleCov.minimum_coverage 99
-SimpleCov.start do
-  add_filter '/spec/'
-end
-
-require 'webmock/rspec'
-
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV['RAILS_ENV'] ||= 'test'
 require File.expand_path('../dummy/config/environment', __FILE__)
@@ -13,6 +5,11 @@ require File.expand_path('../dummy/config/environment', __FILE__)
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'spec_helper'
 require 'rspec/rails'
+require 'simplecov'
+require 'webmock/rspec'
+
+SimpleCov.minimum_coverage 99
+SimpleCov.start 'rails'
 # Add additional requires below this line. Rails is not loaded until this point!
 
 # Requires supporting ruby files with custom matchers and macros, etc, in